Fix casting of byte counts

+++ hoomd/hpmc/PatchEnergyJITGPU.cc
@@ -39,7 +39,7 @@ void PatchEnergyJITGPU::computePatchEnergyGPU(const gpu_args_t& args, hipStream_
 
     unsigned int max_queue_size = n_groups * tpp;
 
-    const size_t min_shared_bytes = args.num_types * sizeof(Scalar);
+    const unsigned int min_shared_bytes = args.num_types * sizeof(Scalar);
 
     unsigned int shared_bytes
         = n_groups
